bug/TS-34: Fix HTTP to support > 2GB document sizes.
Tested: FC12, Ubuntu-9, Ubuntu-10, FBSD.

Make "core" HttpSM and logging support 64-bit "documents". Also
eliminates LOG_INT. There are still areas in logging that needs
"tweaking", but that will be a second bug/patch.

Additional fixes from George, who also reviewed.

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@948906 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/ClusterCache.cc
@@ -1489,7 +1489,7 @@ CacheContinuation::setupVCdataRead(int event, VConnection * vc)
     ink_release_assert(caller_buf_freebytes);
     SET_HANDLER((CacheContHandler) & CacheContinuation::VCdataRead);
 
-    int size_index = iobuffer_size_to_index(caller_buf_freebytes);
+    int64 size_index = iobuffer_size_to_index(caller_buf_freebytes);
     MIOBuffer *buf = new_MIOBuffer(size_index);
     readahead_reader = buf->alloc_reader();
