+++ iocore/cluster/ClusterHandler.cc
@@ -592,7 +592,7 @@ bool ClusterHandler::build_initial_vector(bool read_flag)
               //////////////////////////////////////////////////////////
               ink_release_assert(s.msg.descriptor[i].length <= DEFAULT_MAX_BUFFER_SIZE);
               vc->read_block = new_IOBufferBlock();
-              int index = buffer_size_to_index(s.msg.descriptor[i].length, MAX_BUFFER_SIZE_INDEX);
+              int64 index = buffer_size_to_index(s.msg.descriptor[i].length, MAX_BUFFER_SIZE_INDEX);
               vc->read_block->alloc(index);
 
               s.iov[new_n_iov].iov_base = 0;
